Replace admin-heal-list with admin-heal --fake
